FossilOrigin-Name: 4fae0e1e4597dc3a183d096cb2aea8ee28ce5070b685105fe442445a8ddccd2b

+++ src/www/admin/config/ext/delete.php
@@ -13,6 +13,10 @@ $mode = qg('mode');
 if (qg('plugin')) {
 	$plugin = Plugins::get(qg('plugin'));
 
+	if ($plugin->enabled) {
+		throw new UserException('Impossible de supprimer une extension activée');
+	}
+
 	$form->runIf(f('delete') && f('confirm_delete'), function () use ($plugin) {
 		$plugin->delete();
 	}, $csrf_key, '!config/ext/');
@@ -20,11 +24,21 @@ if (qg('plugin')) {
 else {
 	$module = Modules::get(qg('module'));
 
+	if ($mode === 'data' && !$module->canDeleteData()) {
+		throw new UserException('Impossible de supprimer les données de ce module.');
+	}
+	elseif ($mode === 'reset' && !$module->canReset()) {
+		throw new UserException('Impossible de remettre ce module à son état antérieur.');
+	}
+	elseif ($mode === 'delete' && !$module->canDelete()) {
+		throw new UserException('Impossible de supprimer ce module.');
+	}
+
 	$form->runIf(f('delete') && f('confirm_delete'), function () use ($module, $mode) {
-		if ($mode == 'data') {
+		if ($mode === 'data') {
 			$module->deleteData();
 		}
-		elseif ($mode == 'reset') {
+		elseif ($mode === 'reset') {
 			$module->resetChanges();
 		}
 		else {
